requirements.txt
langchain-core
fastapi
langchain
streamlit
google-generativeai
PyPDF2
scipy
langchain-community
openai